pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "web-security-agent"
version = "1.0.0"
description = "CLI-based Website Security Analysis & Vulnerability Agent with Human-in-the-Loop"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Security Agent Team"}]

dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "jinja2>=3.1",
    "markdown>=3.5",
    "anthropic>=0.40",
    "openai>=1.0",
    "pyyaml>=6.0",
    "validators>=0.22",
    "dnspython>=2.4",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "respx>=0.21",
]

[project.scripts]
secagent = "security_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
